+++ Runtime/ResourceManager/ResourceProviders/BundledAssetProvider.cs
@@ -44,6 +44,19 @@ namespace UnityEngine.ResourceManagement.ResourceProviders
                 return bundle as T;
             }
 
+            internal static bool IsDownloadOnly(IList<object> results)
+            {
+                foreach (var dep in results)
+                {
+                    if (dep is AssetBundleResource { m_DownloadOnly: true })
+                    {
+                        return true;
+                    }
+                }
+
+                return false;
+            }
+
             public void Start(ProvideHandle provideHandle)
             {
                 provideHandle.SetProgressCallback(ProgressCallback);
@@ -63,7 +76,14 @@ namespace UnityEngine.ResourceManagement.ResourceProviders
                     m_AssetBundle = bundleResource.GetAssetBundle();
                     if (m_AssetBundle == null)
                     {
-                        m_ProvideHandle.Complete<AssetBundle>(null, false, new Exception("Unable to load dependent bundle from location " + m_ProvideHandle.Location));
+#if UNITY_EDITOR
+                        // This more detailed error is only displayed in the Editor to help with debugging. It's not
+                        // an Editor only issue.
+                        if (IsDownloadOnly(deps))
+                            m_ProvideHandle.Complete<AssetBundle>(null, false, new Exception("Unable to load dependent bundle from location " + m_ProvideHandle.Location + ".\nRelease the handle returned by DownloadDependenciesAsync or call it with autoReleaseHandle set to true."));
+                        else
+#endif
+                            m_ProvideHandle.Complete<AssetBundle>(null, false, new Exception("Unable to load dependent bundle from location " + m_ProvideHandle.Location));
                         return;
                     }